+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
 ---
 
  * Add `service_closure()` to the PHP-DSL
+ * Add support for autoconfigurable attributes on methods, properties and parameters
 
 5.3
 ---